+++ CHANGELOG.md
@@ -5,4 +5,8 @@ The sections should follow the order `Packaging`, `Added`, `Changed`, `Fixed` an
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-# Unreleased
+## Unreleased
+
+### Added
+
+- Access to alternative names for cursor icons through the `CursorIcon::alt_names` method.